+++ src/collections/source_component.rs
@@ -8,9 +8,12 @@ use enum_dispatch::enum_dispatch;
 use crate::{
     base::{Float, Pose, Source, Transform},
     collections::{SourceArray, SourceAssembly},
-    currents::{CircularCurrent, Current},
+    currents::{CircularCurrent, Current, PathCurrent, TriangleCurrent},
     magnets::{CuboidMagnet, CylinderMagnet, Dipole, Magnet, SphereMagnet},
 };
+
+#[cfg(feature = "mesh")]
+use crate::currents::SheetCurrent;
 use nalgebra::{Point3, Vector3};
 
 #[derive(Debug, Clone)]
@@ -91,7 +94,10 @@ macro_rules! impl_transitive_from_current {
 }
 
 impl_transitive_from_magnet!(CylinderMagnet, CuboidMagnet, Dipole, SphereMagnet);
-impl_transitive_from_current!(CircularCurrent);
+impl_transitive_from_current!(CircularCurrent, PathCurrent, TriangleCurrent);
+
+#[cfg(feature = "mesh")]
+impl_transitive_from_current!(SheetCurrent);
 
 impl<T: Float> Eq for SourceComponent<T> {}