+++ src/currents/current.rs
@@ -9,7 +9,10 @@ use derive_more::Display;
 use enum_dispatch::enum_dispatch;
 use nalgebra::{Point3, Vector3};
 
-use super::{CircularCurrent, PathCurrent};
+use super::{CircularCurrent, PathCurrent, TriangleCurrent};
+
+#[cfg(feature = "mesh")]
+use super::SheetCurrent;
 use crate::base::{Float, Pose, Source, Transform};
 
 /// Current source variants.
@@ -19,4 +22,7 @@ use crate::base::{Float, Pose, Source, Transform};
 pub enum Current<T: Float = f64> {
     Circular(CircularCurrent<T>),
     Path(PathCurrent<T>),
+    #[cfg(feature = "mesh")]
+    Sheet(SheetCurrent<T>),
+    Triangle(TriangleCurrent<T>),
 }